Refactor Vulkan app into modules
