+++ src/vulkan_app/commands.rs
@@ -1,0 +1,95 @@
+use ash::{vk};
+
+use super::{utils::QueueFamilyIndices, vertex::{INDICES}, VulkanApp};
+
+pub fn create_command_pool(device: &ash::Device, indices: &QueueFamilyIndices) -> vk::CommandPool {
+    let pool_info = vk::CommandPoolCreateInfo::builder()
+        .queue_family_index(indices.graphics_family.unwrap())
+        .flags(vk::CommandPoolCreateFlags::empty());
+    unsafe { device.create_command_pool(&pool_info, None).unwrap() }
+}
+
+pub fn create_command_buffers(
+    device: &ash::Device,
+    command_pool: vk::CommandPool,
+    framebuffer_count: usize,
+) -> Vec<vk::CommandBuffer> {
+    let alloc_info = vk::CommandBufferAllocateInfo::builder()
+        .command_pool(command_pool)
+        .level(vk::CommandBufferLevel::PRIMARY)
+        .command_buffer_count(framebuffer_count as u32);
+    unsafe { device.allocate_command_buffers(&alloc_info).unwrap() }
+}
+
+impl VulkanApp {
+    pub fn record_command_buffer(&self, command_buffer: vk::CommandBuffer, image_index: usize) {
+        let begin_info = vk::CommandBufferBeginInfo::builder();
+        unsafe {
+            self.device
+                .begin_command_buffer(command_buffer, &begin_info)
+                .unwrap();
+        }
+
+        let clear_color = vk::ClearValue {
+            color: vk::ClearColorValue { float32: [0.0, 0.0, 0.0, 1.0] },
+        };
+        let depth_clear = vk::ClearValue {
+            depth_stencil: vk::ClearDepthStencilValue { depth: 1.0, stencil: 0 },
+        };
+        let clear_values = [clear_color, depth_clear];
+        let render_pass_info = vk::RenderPassBeginInfo::builder()
+            .render_pass(self.render_pass)
+            .framebuffer(self.framebuffers[image_index])
+            .render_area(vk::Rect2D {
+                offset: vk::Offset2D { x: 0, y: 0 },
+                extent: self.swapchain_extent,
+            })
+            .clear_values(&clear_values);
+
+        unsafe {
+            self.device.cmd_begin_render_pass(
+                command_buffer,
+                &render_pass_info,
+                vk::SubpassContents::INLINE,
+            );
+            self.device.cmd_bind_pipeline(
+                command_buffer,
+                vk::PipelineBindPoint::GRAPHICS,
+                self.graphics_pipeline,
+            );
+            let vertex_buffers = [self.vertex_buffer];
+            let offsets = [0];
+            self.device
+                .cmd_bind_vertex_buffers(command_buffer, 0, &vertex_buffers, &offsets);
+            self.device.cmd_bind_index_buffer(
+                command_buffer,
+                self.index_buffer,
+                0,
+                vk::IndexType::UINT16,
+            );
+            self.device.cmd_bind_descriptor_sets(
+                command_buffer,
+                vk::PipelineBindPoint::GRAPHICS,
+                self.pipeline_layout,
+                0,
+                &[self.descriptor_sets[image_index]],
+                &[],
+            );
+            self.device
+                .cmd_draw_indexed(command_buffer, INDICES.len() as u32, 1, 0, 0, 0);
+            self.device.cmd_end_render_pass(command_buffer);
+            self.device.end_command_buffer(command_buffer).unwrap();
+        }
+    }
+}
+
+pub fn create_sync_objects(device: &ash::Device) -> (vk::Semaphore, vk::Semaphore, vk::Fence) {
+    let semaphore_info = vk::SemaphoreCreateInfo::builder();
+    let fence_info = vk::FenceCreateInfo::builder().flags(vk::FenceCreateFlags::SIGNALED);
+
+    let image_available_semaphore = unsafe { device.create_semaphore(&semaphore_info, None).unwrap() };
+    let render_finished_semaphore = unsafe { device.create_semaphore(&semaphore_info, None).unwrap() };
+    let in_flight_fence = unsafe { device.create_fence(&fence_info, None).unwrap() };
+
+    (image_available_semaphore, render_finished_semaphore, in_flight_fence)
+}
